feat(auth)!: remove api auth and include functionality in middleware
